refactor: Separate BIOS Setup Utility from boot screen logic
- Moved CMOS Setup Utility logic (TUI drawing and interaction) to a dedicated `src/system/setup-utility.js` file.
- Cleaned up `src/system/boot-screen.js` by importing the setup TUI logic.
- Improved modularity by using callbacks for setup exit.
